refactor(events): extract core::events into headless ts-events crate (#91)
Вынос event-driven фундамента (AppEvent / EventBus / EventHandler) из монолита
src-tauri/src/core/events.rs в новый headless-крейт ts-events. Это первый шаг к
разблокировке группы команд ai_api_proxy (использует AppEvent), а также самостоятельная
foundational-инфраструктура (эпик #91 хочет EventBus отдельным крейтом).

- crates/ts-events: новый крейт (serde/specta/tokio/async-trait + ts-render ради типа
  ошибки Result/VideoCompilerError). Добавлен в workspace members.
- git mv core/events.rs -> crates/ts-events/src/lib.rs; переписаны пути:
  crate::video_compiler::error -> ts_render::video_compiler::error;
  в макросе event_handler!: $crate::core::EventHandler -> $crate::EventHandler,
  $crate::video_compiler::error::Result -> ts_render::...::Result.
- Монолит: core/events.rs стал ре-экспорт-шимом `pub use ts_events::*;` —
  core/mod.rs::`pub use events::{AppEvent, EventBus, EventHandler}` и ~7 потребителей
  резолвятся без правок. + path-dep ts-events в src-tauri/Cargo.toml.

specta-фичи крейта синхронны с монолитом (serde/chrono/uuid/serde_json) — AppEvent
деривит specta::Type с полями serde_json::Value / DateTime<Utc>.

Проверки: cargo test -p ts-events -> 24 passed; cargo check --lib (src-tauri) -> 0 ошибок,
3 baseline-варнинга.
