+++ src-tauri/Cargo.toml
@@ -90,6 +90,7 @@ ts-state = { path = "../crates/ts-state", features = ["specta"] }
 ts-agent = { path = "../crates/ts-agent" }
 ts-platform = { path = "../crates/ts-platform" }
 ts-analysis = { path = "../crates/ts-analysis" }
+ts-events = { path = "../crates/ts-events" }
 ts-render = { path = "../crates/ts-render" }
 ts-render-services = { path = "../crates/ts-render-services" }
 ts-render-tauri = { path = "../crates/ts-render-tauri" }
